Use uint64_t instead of ugly unsigned long long...

+++ src/core/packet.h
@@ -67,7 +67,7 @@ struct packet_info {
 	 */
 	size_t hdr_length;
 	// Packet timestamp in microseconds since epoch
-	unsigned long long int timestamp;
+	uint64_t timestamp;
 	/*
 	 * Source and destination address. Raw data (addr_len bytes each).
 	 * Is set only with ip_protocol == 4 || 6, or with ethernet frames.
